feat: token login, email password login, hop ignite rollout

+++ src/commands/ignite/types.rs
@@ -78,7 +78,7 @@ impl RamSizes {
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Resources {
-    pub vcpu: u64,
+    pub vcpu: f64,
     pub ram: String,
     #[serde(skip)]
     pub vgpu: Vec<Vgpu>,
@@ -87,7 +87,7 @@ pub struct Resources {
 impl Default for Resources {
     fn default() -> Self {
         Resources {
-            vcpu: 1,
+            vcpu: 1.,
             ram: RamSizes::default().to_string(),
             vgpu: vec![],
         }
